created a new post method for the email templates that autp associates the the creation with the user

+++ frontend/src/ReduxActionsMain/myEmailTemplateAction.js
@@ -1,3 +1,4 @@
+const token = localStorage.getItem('jwt)');
 
 export const setSelectedMyEmail = (file) => { 
     console.log(file)
@@ -10,9 +11,17 @@ export const setSelectedMyEmail = (file) => {
   
 // Actin to fetch my email templates 
   export const fetchMyEmails = () => {
+    if (!token) {
+      throw new Error("Authorization token is missing.");
+    }
     return async (dispatch) => {
       try {
-        const response = await fetch('http://127.0.0.1:3000/my_templates');
+        const response = await fetch('http://127.0.0.1:3000/my_templates', {
+          headers: { 
+            "Content-Type": "application/json", 
+            Authorization: `Bearer ${token}`
+           },
+        });
         const data = await response.json();
         dispatch({ type: 'FETCH_MY_EMAILS_SUCCESS', payload: data });
       } catch (error) {
@@ -24,11 +33,15 @@ export const setSelectedMyEmail = (file) => {
   // Action to create a patient email template
   export const createMyEmail = (newEmail) => {
     return async (dispatch) => {
+      if (!token) {
+        throw new Error("Authorization token is missing.");
+      }
       try {
         const response = await fetch('http://127.0.0.1:3000/my_templates', {
           method: 'POST',
           headers: {
             'Content-Type': 'application/json',
+            Authorization: `Bearer ${token}`
           },
           body: JSON.stringify(newEmail),
         });
@@ -62,9 +75,13 @@ export const setSelectedMyEmail = (file) => {
   // Action to delete a patient email template
   export const deletePatientEmail = (emailId) => {
     return async (dispatch) => {
+      if (!token) {
+        throw new Error("Authorization token is missing.");
+      }
       try {
         await fetch(`http://127.0.0.1:3000/my_templates/${emailId}`, {
           method: 'DELETE',
+          Authorization: `Bearer ${token}`
         });
         dispatch({ type: 'DELETE__MY_EMAIL_SUCCESS', payload: emailId });
       } catch (error) {
